backend/requirements.txt
azure-functions
azure-storage-blob
pymongo
openai
azure-ai-documentintelligence
langchain-text-splitters
firebase-admin
